feat: organized components, CSS into proper directories
