+++ src/CSS/AboutMe.css
@@ -54,3 +54,19 @@
     color:white;
     padding:1rem;
 }
+
+@media (max-width: 48rem) {
+    .img-and-p-wrapper {
+        width: 90%;
+    }
+
+    .img-and-p-wrapper img {
+        height: 12rem;
+        max-width: 100%;
+        object-fit: cover;
+    }
+
+    .img-and-p-wrapper p {
+        padding: 1rem;
+    }
+}
