+++ src/CSS/App.css
@@ -1,6 +1,6 @@
 #root {
-  height:100vh;
-  width:100vw; 
+  min-height:100vh;
+  width:100%;
   margin: 0 auto;
   text-align: center;
 }
