+++ src/CSS/Contact.css
@@ -142,3 +142,30 @@ input {
 a {
     font-size: x-large;
 }
+
+@media (max-width: 48rem) {
+    .contact-wrapper {
+        margin: 1rem;
+    }
+
+    .contact-form-wrapper {
+        width: 100%;
+        box-sizing: border-box;
+        padding: 1.5rem;
+    }
+
+    .contact-form-wrapper form {
+        width: 100%;
+        max-width: 28rem;
+    }
+
+    .info-section {
+        flex-wrap: wrap;
+    }
+
+    .person {
+        flex-wrap: wrap;
+        justify-content: center;
+        margin: 1rem;
+    }
+}
