+++ src/CSS/Header.css
@@ -1,7 +1,8 @@
 .header {
   display: flex;
   flex-direction: column;
-  width: 100vw;
+  width: 100%;
+  box-sizing: border-box;
   background-color: #2936c0;
   padding: 1rem;
   border: .2rem solid white;
@@ -29,4 +30,20 @@
 .nav-links a:hover {
   text-decoration: underline;
   color: orange;
+}
+
+@media (max-width: 30rem) {
+  .header {
+    padding: 0.75rem 0.5rem;
+  }
+
+  .nav-links {
+    flex-wrap: wrap;
+    gap: 0.5rem 1rem;
+    justify-content: center;
+  }
+
+  .nav-links a {
+    font-size: 0.9rem;
+  }
 }
