+++ src/Components/Header.jsx
@@ -1,5 +1,5 @@
 import {Link} from 'react-router-dom'
-import './Header.css'
+import '../CSS/Header.css'
 
 const Header = () => {
     return (
