+++ src/Components/Resume.jsx
@@ -1,4 +1,4 @@
-import './Resume.css'
+import '../CSS/Resume.css'
 
 const Resume = () => {
     return (
